backend/requirements.txt
# Python dependencies for AURA Voice AI

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# API clients
openai==1.3.0
httpx==0.25.0

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4

# Memory and caching
redis[hiredis]==5.0.1
asyncio==3.4.3

# Voice processing
elevenlabs==0.2.26

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1
aiofiles==23.2.1

# AI and ML
tiktoken==0.5.1
numpy==1.24.3
scikit-learn==1.3.2

# Utilities
python-dateutil==2.8.2
beautifulsoup4==4.12.2
chardet==5.2.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0